+++ examples/push_button.rs
@@ -1,0 +1,35 @@
+extern crate pushrod;
+extern crate sdl2;
+
+use pushrod::render::engine::Engine;
+use pushrod::render::widget::Widget;
+use pushrod::render::widget_config::{CONFIG_BORDER_WIDTH, CONFIG_COLOR_BASE, CONFIG_COLOR_BORDER};
+use pushrod::widgets::push_button_widget::PushButtonWidget;
+use sdl2::pixels::Color;
+
+/*
+ * This demo just tests the rendering functionality of the `BaseWidget`.  It only tests the
+ * render portion of the library, nothing else.
+ */
+
+pub fn main() {
+    let sdl_context = sdl2::init().unwrap();
+    let video_subsystem = sdl_context.video().unwrap();
+    let window = video_subsystem
+        .window("pushrod-render demo", 800, 600)
+        .position_centered()
+        .opengl()
+        .build()
+        .unwrap();
+    let mut engine = Engine::new();
+    let mut button1 = PushButtonWidget::new(16, 16, 80, 48);
+
+    button1.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
+    button1.set_numeric(CONFIG_BORDER_WIDTH, 2);
+
+    engine.setup(800, 600);
+
+    engine.add_widget(Box::new(button1), String::from("button1"));
+
+    engine.run(sdl_context, window);
+}