+++ src/render/widget.rs
@@ -140,34 +140,46 @@ pub trait Widget {
     ) {
     }
 
+    /// This callback is called when a setter is used to configure a value.  It is _not_ called when a
+    /// call to `get_config()` using the setter is called, so it is best to use the top-level setters
+    /// and getters for the configuration values - at least, until the `get_config()` call can be made
+    /// private.
+    fn on_config_changed(&mut self, _k: u8, _v: Config) {}
+
     /// Sets a point for a configuration key.
     fn set_point(&mut self, config: u8, x: i32, y: i32) {
         self.get_config().set_point(config, x, y);
+        self.on_config_changed(config, Config::Points(vec![x, y]));
     }
 
     /// Sets a color for a configuration key.
     fn set_color(&mut self, config: u8, color: Color) {
         self.get_config().set_color(config, color);
+        self.on_config_changed(config, Config::Color(color));
     }
 
     /// Sets a numeric value for a configuration key.
     fn set_numeric(&mut self, config: u8, value: i32) {
         self.get_config().set_numeric(config, value);
+        self.on_config_changed(config, Config::Numeric(value));
     }
 
     /// Sets a text value for a configuration key.
     fn set_text(&mut self, config: u8, text: String) {
-        self.get_config().set_text(config, text);
+        self.get_config().set_text(config, text.clone());
+        self.on_config_changed(config, Config::Text(text.clone()));
     }
 
     /// Sets a toggle for a configuration key.
     fn set_toggle(&mut self, config: u8, flag: bool) {
         self.get_config().set_toggle(config, flag);
+        self.on_config_changed(config, Config::Toggle(flag));
     }
 
     /// Sets a compass position for a configuration key.
     fn set_compass(&mut self, config: u8, value: CompassPosition) {
-        self.get_config().set_compass(config, value);
+        self.get_config().set_compass(config, value.clone());
+        self.on_config_changed(config, Config::CompassPosition(value.clone()));
     }
 
     /// Retrieves a `Points` for a configuration key.  Returns `Points::default` if not set.