+++ src/widgets/progress_widget.rs
@@ -83,10 +83,11 @@ impl Widget for ProgressWidget {
         .unwrap();
     }
 
-    /// Sets a numeric value for a configuration key.
-    fn set_numeric(&mut self, config: u8, value: i32) {
-        self.get_config().set_numeric(config, value);
-        self.get_config().set_invalidate(true);
+    /// Responds to a screen redraw only if the `CONFIG_PROGRESS` key was changed.
+    fn on_config_changed(&mut self, _k: u8, _v: Config) {
+        if _k == CONFIG_PROGRESS {
+            self.get_config().set_invalidate(true);
+        }
     }
 
     default_widget_properties!();