+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Updated documentation in Timer.
 - Modified `button_clicked` function in `engine` so that when a mouse is released, it is sent to all `Widget`s.
 - Modified `button_clicked` function in `WidgetCache` so that a `widget_id == -1` sends a message to all `Widget`s.
+- Added `on_config_changed` so that it is called when a config option changes (#215)
+- Modified `ProgressWidget` and `ImageWidget` to use `on_config_changed` to handle better redraw.
 
 ## 0.4.9